Make jb-rotary service more robust
Modified jb-rotary.py to detect the justboom card even when the
on-board audio (ALSA) is active.

Not sure if the possible card names have to be updated/extended.

Had to delay the start of the service, since it started too early
before the alsa devices were available. Current delay is 20 sec. Hope
that is sufficient for all cases.